sources/apps/AlphaBot/requirements.txt
paho-mqtt
RPi.GPIO; platform_machine == "armv7l" or platform_machine == "aarch64"
rpi_ws281x; platform_machine == "armv7l" or platform_machine == "aarch64"